pyproject.toml
[tool.poetry]
name = "dans-file-format-service"
version = "0.1.0"
description = ""
authors = ["Eko Indarto <eko.indarto@dans.knaw.nl>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
dynaconf = "^3.1.11"
requests = "^2.28.1"
python-magic = "^0.4.27"
puremagic = "^1.14"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
